[Avani] Green 1.2- should return total cart when add one apple.

+++ src/test/java/com/thoughtworks/bootcamp/ShoppingCartTest.java
@@ -1,12 +1,15 @@
 package com.thoughtworks.bootcamp;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class ShoppingCartTest {
     @Test
     void givenEmptyShoppingCart_whenAddOneApple_thenShouldReturnOneApple() {
         ShoppingCart shoppingCart = new ShoppingCart("apple", 1, 0.99);
-        float cartTotal = shoppingCart.calculateCartTotal();
+
+        double cartTotal = shoppingCart.calculateCartTotal();
+
         Assertions.assertEquals(0.99, cartTotal);
 
     }
